+++ data/duke.txt
@@ -2,3 +2,7 @@
 [E][X] movie night (at: 23 Aug 2021)
 [D][ ] cs2103t (by: 25 Aug 2021 23:59)
 [T][ ] finish sogp stuff
+[T][ ] testing
+[D][ ] cs2013t (by: 15:00)
+[D][ ] alloc (by: 12:00)
+[E][ ] breakfast (at: 09:15)
