+++ src/main/java/duke/tasks/Deadline.java
@@ -1,6 +1,11 @@
+package duke.tasks;
+
+import duke.DukeException;
+
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
+import java.time.format.DateTimeParseException;
 
 /**
  * Represents a deadline, which is a subtype of a Task.
@@ -16,15 +21,20 @@ public class Deadline extends Task {
      * @param description Description of the Deadline.
      * @param by The date and/or time by which the task should be done.
      */
-    public Deadline(String description, String by) {
+    public Deadline(String description, String by) throws DukeException {
         super(description);
-        if (by.contains(" ")) {
-            this.date = LocalDate.parse(by.split(" ", 2)[0]);
-            this.time = LocalTime.parse(by.split(" ", 2)[1]);
-        } else if (by.contains("-")) {
-            this.date = LocalDate.parse(by);
-        } else {
-            this.time = LocalTime.parse(by);
+        try {
+            if (by.contains(" ")) {
+                this.date = LocalDate.parse(by.split(" ", 2)[0]);
+                this.time = LocalTime.parse(by.split(" ", 2)[1]);
+            } else if (by.contains("-")) {
+                this.date = LocalDate.parse(by);
+            } else {
+                this.time = LocalTime.parse(by);
+            }
+        }
+        catch (DateTimeParseException e) {
+            throw new DukeException("Use the right format for date and/or time: yyyy-MM-dd and/or HH:mm");
         }
     }
 
@@ -35,9 +45,9 @@ public class Deadline extends Task {
      */
     @Override
     public String toString() {
-        String toPrint = "[D]" + super.toString() + " (by: ";
+        String toPrint = "[D]" + super.toString() + " (by:";
         if (date != null) {
-            toPrint += date.format(DateTimeFormatter.ofPattern("dd MMM yyyy"));
+            toPrint += " " + date.format(DateTimeFormatter.ofPattern("dd MMM yyyy"));
         }
         if (time != null) {
             toPrint += " " + time.format(DateTimeFormatter.ofPattern("HH:mm"));
